server/agent/memory/vector_store.py
from __future__ import annotations
"""
Stub for pgvector integration. Not used in minimal graph yet.
Add table schemas and embedding upsert/query methods here when adding RAG.
"""
from typing import Any, List, Optional

try:
    from sqlalchemy import create_engine, text
    _SQLA = True
except Exception:
    _SQLA = False

from ...infra.settings import settings


class VectorStore:
    def __init__(self) -> None:
        self.dsn = settings.pg_dsn
        self.engine = create_engine(self.dsn) if (_SQLA and self.dsn) else None

    def ready(self) -> bool:
        return self.engine is not None

    def ensure_schema(self) -> None:
        if not self.engine:
            return
        # Define your pgvector schemas here
        with self.engine.begin() as conn:
            conn.execute(text("""
            CREATE TABLE IF NOT EXISTS kb_chunks (
              id SERIAL PRIMARY KEY,
              doc_id TEXT,
              chunk TEXT,
              embedding vector(1536),
              meta JSONB
            );
            """))

    def search(self, query_emb: List[float], k: int = 5) -> List[dict[str, Any]]:
        if not self.engine:
            return []
        # Implement a pgvector ANN search
        return []
