server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
sse-starlette==2.1.3
redis==5.0.7
google-generativeai==0.7.2
